+++ Data/SBV/Control/Utils.hs
@@ -1310,7 +1310,8 @@ getAllSatResult = do queryDebug ["*** Checking Satisfiability, all solutions.."]
                                       -- Add on observables if we're asked to do so:
                                       obsvs <- if grabObservables
                                                   then getObservables
-                                                  else return []
+                                                  else do queryDebug ["*** In a quantified context, obvservables will not be printed."]
+                                                          return []
 
                                       bindings <- let grab i@(ALL, _)      = return (i, Nothing)
                                                       grab i@(EX, (sv, _)) = case sv `lookup` assocs of